+++ src/main/java/mekanism/client/render/RenderResizableCuboid.java
@@ -129,7 +129,7 @@ public class RenderResizableCuboid {
      */
     public void renderCubeFromCentre(Model3D cuboid) {
         GlStateManager.pushMatrix();
-        GlStateManager.translate(-cuboid.sizeX() / 2D, -cuboid.sizeY() / 2D, -cuboid.sizeZ() / 2D);
+        GlStateManager.translate((float) -cuboid.sizeX() / 2F, (float) -cuboid.sizeY() / 2F, (float) -cuboid.sizeZ() / 2F);
         renderCube(cuboid, EnumShadeArgument.NONE, null, null, null);
         GlStateManager.popMatrix();
     }
