+++ src/main/java/mekanism/generators/client/render/RenderHeatGenerator.java
@@ -18,7 +18,7 @@ public class RenderHeatGenerator extends TileEntitySpecialRenderer<TileEntityHea
     @Override
     public void render(TileEntityHeatGenerator tileEntity, double x, double y, double z, float partialTick, int destroyStage, float alpha) {
         GlStateManager.pushMatrix();
-        GlStateManager.translate(x + 0.5, y + 1.5, z + 0.5);
+        GlStateManager.translate((float) x + 0.5F, (float) y + 1.5F, (float) z + 0.5F);
         bindTexture(MekanismUtils.getResource(ResourceType.RENDER, "HeatGenerator.png"));
 
         GLSMHelper.rotate(tileEntity.facing, 180, 0, 270, 90);
